refactor block_utils.js
