+++ apps/src/block_utils.js
@@ -1063,67 +1063,20 @@ exports.createJsWrapperBlockCreator = function (
           this.setPreviousStatement(true);
         }
 
+        // Boolean constant to store when we show mini-toolbox.
         // Use window.appOptions, not global appOptions, because the levelbuilder
         // block page doesn't have appOptions, but we *do* want to show the mini-toolbox
-        // there
-        if (
+        // there.
+        const showMiniToolbox =
           miniToolboxBlocks &&
-          (!window.appOptions || window.appOptions.level.miniToolbox)
-        ) {
-          var toggle = new Blockly.FieldIcon('+');
-          if (Blockly.cdoUtils.isWorkspaceReadOnly(this.blockSpace)) {
-            toggle.setReadOnly();
-          }
+          (!window.appOptions || window.appOptions.level.miniToolbox);
 
-          var miniToolboxXml = '<xml>';
-          miniToolboxBlocks.forEach(block => {
-            miniToolboxXml += `\n <block type="${block}"></block>`;
-          });
-          miniToolboxXml += '\n</xml>';
-          // Block.isMiniFlyoutOpen is used in the blockly repo to track whether or not the horizontal flyout is open.
-          this.isMiniFlyoutOpen = false;
-          // On button click, open/close the horizontal flyout, toggle button text between +/-, and re-render the block.
-          Blockly.cdoUtils.bindBrowserEvent(
-            toggle.fieldGroup_,
-            'mousedown',
-            this,
-            () => {
-              if (Blockly.cdoUtils.isWorkspaceReadOnly(this.blockSpace)) {
-                return;
-              }
-
-              if (this.isMiniFlyoutOpen) {
-                toggle.setValue('+');
-              } else {
-                toggle.setValue('-');
-              }
-              this.isMiniFlyoutOpen = !this.isMiniFlyoutOpen;
-              this.render();
-              // If the mini flyout just opened, make sure mini-toolbox blocks are updated with the right thumbnails.
-              // This has to happen after render() because some browsers don't render properly if the elements are not
-              // visible. The root cause is that getComputedTextLength returns 0 if a text element is not visible, so
-              // the thumbnail image overlaps the label in Firefox, Edge, and IE.
-              if (this.isMiniFlyoutOpen) {
-                let miniToolboxBlocks = this.miniFlyout.blockSpace_.topBlocks_;
-                let rootInputBlocks = this.getConnections_(true /* all */)
-                  .filter(function (connection) {
-                    return connection.type === Blockly.INPUT_VALUE;
-                  })
-                  .map(function (connection) {
-                    return connection.targetBlock();
-                  });
-                miniToolboxBlocks.forEach(function (block, index) {
-                  block.shadowBlockValue_(rootInputBlocks[index]);
-                });
-              }
-            }
-          );
-
-          this.appendDummyInput()
-            .appendField(toggle, 'toggle')
-            .appendField(' ');
-
-          this.initMiniFlyout(miniToolboxXml);
+        let flyoutToggleButton;
+        if (showMiniToolbox) {
+          flyoutToggleButton =
+            Blockly.customBlocks.initializeMiniToolbox.bind(this)(
+              miniToolboxBlocks
+            );
         }
 
         // These blocks should not be loaded into a Google Blockly level.
@@ -1170,6 +1123,13 @@ exports.createJsWrapperBlockCreator = function (
         }
         interpolateInputs(blockly, this, inputRows, inputTypes, inline);
         this.setInputsInline(inline);
+
+        if (showMiniToolbox) {
+          Blockly.customBlocks.initializeFlyoutToggle.bind(this)(
+            miniToolboxBlocks,
+            flyoutToggleButton
+          );
+        }
       },
     };
