+++ apps/src/blockly/cdoBlocklyWrapper.js
@@ -1,6 +1,7 @@
 import {BlocklyVersion} from '@cdo/apps/blockly/constants';
 import {CLAMPED_NUMBER_REGEX} from './constants';
 import {APP_HEIGHT} from '@cdo/apps/p5lab/constants';
+import customBlocks from './customBlocks/cdoBlockly/index.js';
 
 const INFINITE_LOOP_TRAP =
   '  executionInfo.checkTimeout(); if (executionInfo.isTerminated()){return;}\n';
@@ -318,6 +319,7 @@ function initializeBlocklyWrapper(blocklyInstance) {
       return blockSpace.blockSpaceEditor.svgResize();
     },
   };
+  blocklyWrapper.customBlocks = customBlocks;
   return blocklyWrapper;
 }
