+++ apps/src/blockly/customBlocks/cdoBlockly/index.js
@@ -1,0 +1,5 @@
+import {blocks as spritelabBlocks} from './spritelabBlocks';
+
+export default {
+  ...spritelabBlocks,
+};
