+++ apps/src/blockly/customBlocks/googleBlockly/spritelabBlocks.js
@@ -1,0 +1,3 @@
+export const blocks = {
+  initializeMiniToolbox() {},
+};
